create a new hike; toast messages

+++ app/src/main/java/com/example/randoexpress/RandoListAdapter.kt
@@ -36,7 +36,7 @@ class RandoListAdapter(val data: ArrayList<Model.Rando>, val destination: Int) :
         holder.location.text = data[position].ville
         holder.dateTime.text = "${data[position].heureDepart} ${data[position].dateDepart}"
         val bundle = Bundle()
-        bundle.putInt("randoId", data[position].id)
+        bundle.putInt("randoId", data[position].id as Int)
         bundle.putInt("randoOrigin", destination)
         holder.itemView.setOnClickListener { view ->
             Navigation
